libpressio

Enum PressioOption

#[non_exhaustive]
pub enum PressioOption {

    bool(Option<bool>),
    int8(Option<i8>),
    int16(Option<i16>),
    int32(Option<i32>),
    int64(Option<i64>),
    uint8(Option<u8>),
    uint16(Option<u16>),
    uint32(Option<u32>),
    uint64(Option<u64>),
    float32(Option<f32>),
    float64(Option<f64>),
    string(Option<String>),
    vec_string(Option<Vec<String>>),
    data(Option<PressioData>),
    user_ptr(Option<*mut c_void>),
    dtype(Option<PressioDtype>),
    thread_safety(Option<PressioThreadSafety>),
    unset,

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

bool(Option<bool>)

int8(Option<i8>)

int16(Option<i16>)

int32(Option<i32>)

int64(Option<i64>)

uint8(Option<u8>)

uint16(Option<u16>)

uint32(Option<u32>)

uint64(Option<u64>)

float32(Option<f32>)

float64(Option<f64>)

string(Option<String>)

vec_string(Option<Vec<String>>)

data(Option<PressioData>)

user_ptr(Option<*mut c_void>)

dtype(Option<PressioDtype>)

thread_safety(Option<PressioThreadSafety>)

unset

Implementations

impl PressioOption

pub fn copy_type_only(&self) -> Self

Copy the option type without copying the option value.

The returned PressioOption contains the same variant as self but has None instead of the value that self may have contained.